Trait rust_tracer::tracer::object::Object

pub trait Object: Sync {
    fn hit(&self, r: &Ray) -> Option<Hit<'_>>;
    fn material(&self) -> &Material;
    fn area(&self) -> f64;

    fn normal_at(&self, _p: DVec3) -> DVec3 { ... }
    fn is_translucent(&self) -> bool { ... }
    fn size(&self) -> usize { ... }
    fn inside(&self, _r: &Ray) -> bool { ... }
    fn sample_towards(&self, _p: DVec3, _rand_sq: DVec2) -> DVec3 { ... }
    fn sample_on(&self, _rand_sq: DVec2) -> DVec3 { ... }
    fn sample_from(&self, rand_sq_o: DVec2, rand_sq_d: DVec2) -> Ray { ... }
    fn sample_towards_pdf(&self, p: DVec3, dir: DVec3, h: &Hit<'_>) -> f64 { ... }
}

Common functionality shared between all objects.

Required Methods

fn hit(&self, r: &Ray) -> Option<Hit<'_>>

Does the ray hit the object?

fn material(&self) -> &Material

fn area(&self) -> f64

Surface area of the object

Provided Methods

fn normal_at(&self, _p: DVec3) -> DVec3

Normal of the object at the point p. Does not check if p is actually on the object.

fn is_translucent(&self) -> bool

fn size(&self) -> usize

Number of objects the object consists of.

fn inside(&self, _r: &Ray) -> bool

Is the ray inside the object?

fn sample_towards(&self, _p: DVec3, _rand_sq: DVec2) -> DVec3

Sample a random point from surface of the object that is visible from p Returns direction from p to us. DO THIS BETTER

fn sample_on(&self, _rand_sq: DVec2) -> DVec3

Random point on the surface of the object

fn sample_from(&self, rand_sq_o: DVec2, rand_sq_d: DVec2) -> Ray

Sample random ray leaving the object

fn sample_towards_pdf(&self, p: DVec3, dir: DVec3, h: &Hit<'_>) -> f64

PDF for sampling points on the surface that are visible from p

Arguments

p - The point form which the point sampled on the object should be visible

dir - Direction from p to the sampled point on surface

h - Hit on the surface from p